+++ CHANGELOG.md
@@ -20,8 +20,20 @@ For more information on SemVer, please visit http://semver.org/.
 
 ---
 
-## Version 1.4.4
+## Version 1.4.5
 ### 2013-11-??
+
+#### Improvements
+
+* Improved `wpdk_enqueue_script_page_template()` function
+
+#### Experimental
+
+* Added `WPDKHTML::endJavascriptCompress()`
+
+
+## Version 1.4.4
+### 2013-11-22
 
 #### Bugs